cincuenta: cluster g0 added
